Add Cedar (this refuses to compile because of the doubles)
Based on this example, I do not think that Cedar is currently
a viable candidate for a testing framework in a pure Swift project. Cedar's
Swift support appears to be only with respect to writing Swift Spec files
in an Objective-C project.  To see this mixed mode in action, see:

https://github.com/briancroom/TastesOfCocoa/tree/master/CedarWithNimble